# contracts/openapi/lead-intelligence-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "Company": {
        "additionalProperties": false,
        "description": "Normalized company..",
        "properties": {
          "canonical_name": {
            "title": "Canonical Name",
            "type": "string"
          },
          "country_code": {
            "title": "Country Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "employee_count": {
            "title": "Employee Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "industry": {
            "title": "Industry",
            "type": "string"
          },
          "primary_domain": {
            "title": "Primary Domain",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "canonical_name",
          "primary_domain",
          "country_code",
          "industry",
          "employee_count"
        ],
        "title": "Company",
        "type": "object"
      },
      "CompanyCreate": {
        "additionalProperties": false,
        "description": "Create request.",
        "properties": {
          "canonical_name": {
            "title": "Canonical Name",
            "type": "string"
          },
          "country_code": {
            "title": "Country Code",
            "type": "string"
          },
          "employee_count": {
            "title": "Employee Count",
            "type": "integer"
          },
          "industry": {
            "title": "Industry",
            "type": "string"
          },
          "primary_domain": {
            "title": "Primary Domain",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "workspace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "canonical_name",
          "primary_domain",
          "country_code",
          "industry",
          "employee_count"
        ],
        "title": "CompanyCreate",
        "type": "object"
      },
      "CompanyUpdate": {
        "additionalProperties": false,
        "description": "Partial update request.",
        "properties": {
          "canonical_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Name"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "employee_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Count"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "primary_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Domain"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "workspace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "title": "CompanyUpdate",
        "type": "object"
      },
      "CostLedger": {
        "additionalProperties": false,
        "description": "Tracked cost event..",
        "properties": {
          "campaign_run_id": {
            "format": "uuid",
            "title": "Campaign Run Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "default": "USD",
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "unit_cost": {
            "title": "Unit Cost",
            "type": "number"
          },
          "units": {
            "title": "Units",
            "type": "number"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "campaign_run_id",
          "service",
          "units",
          "unit_cost"
        ],
        "title": "CostLedger",
        "type": "object"
      },
      "CostLedgerCreate": {
        "additionalProperties": false,
        "description": "Create request.",
        "properties": {
          "campaign_run_id": {
            "format": "uuid",
            "title": "Campaign Run Id",
            "type": "string"
          },
          "currency": {
            "default": "USD",
            "title": "Currency",
            "type": "string"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "unit_cost": {
            "title": "Unit Cost",
            "type": "number"
          },
          "units": {
            "title": "Units",
            "type": "number"
          },
          "workspace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "campaign_run_id",
          "service",
          "units",
          "unit_cost"
        ],
        "title": "CostLedgerCreate",
        "type": "object"
      },
      "CostLedgerUpdate": {
        "additionalProperties": false,
        "description": "Partial update request.",
        "properties": {
          "campaign_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Run Id"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "service": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service"
          },
          "unit_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Cost"
          },
          "units": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Units"
          },
          "workspace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "title": "CostLedgerUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "VA Lead Intelligence API",
    "version": "1.0.0",
    "x-build-sha": "unknown",
    "x-generated-at": "2026-07-15T10:45:31Z"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/companies/": {
      "get": {
        "operationId": "list_company_companies__get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Company"
                  },
                  "title": "Response List Company Companies  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Company",
        "tags": [
          "company"
        ]
      },
      "post": {
        "operationId": "create_company_companies__post",
        "parameters": [
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Company",
        "tags": [
          "company"
        ]
      }
    },
    "/companies/{company_id}": {
      "delete": {
        "operationId": "delete_company_companies__company_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Company",
        "tags": [
          "company"
        ]
      },
      "get": {
        "operationId": "get_company_companies__company_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Company",
        "tags": [
          "company"
        ]
      },
      "patch": {
        "operationId": "update_company_companies__company_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Company",
        "tags": [
          "company"
        ]
      }
    },
    "/costs/": {
      "get": {
        "operationId": "list_cost_ledger_costs__get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CostLedger"
                  },
                  "title": "Response List Cost Ledger Costs  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Cost Ledger",
        "tags": [
          "cost_ledger"
        ]
      },
      "post": {
        "operationId": "create_cost_ledger_costs__post",
        "parameters": [
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostLedgerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostLedger"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Cost Ledger",
        "tags": [
          "cost_ledger"
        ]
      }
    },
    "/costs/{cost_ledger_id}": {
      "delete": {
        "operationId": "delete_cost_ledger_costs__cost_ledger_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "cost_ledger_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cost Ledger Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Cost Ledger",
        "tags": [
          "cost_ledger"
        ]
      },
      "get": {
        "operationId": "get_cost_ledger_costs__cost_ledger_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cost_ledger_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cost Ledger Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostLedger"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cost Ledger",
        "tags": [
          "cost_ledger"
        ]
      },
      "patch": {
        "operationId": "update_cost_ledger_costs__cost_ledger_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "cost_ledger_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cost Ledger Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostLedgerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostLedger"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Cost Ledger",
        "tags": [
          "cost_ledger"
        ]
      }
    },
...
  }
}